requirements.txt
arrow==0.17.0
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap-modal-forms==2.0.0
django-chartit==0.2.9
django-filter==2.3.0
django-import-export==2.4.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
pi==0.1.2
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytube==9.7.0
pytube3==9.6.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
times==0.7
typing-extensions==3.7.4.3
urllib3==1.25.11
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
youtube-dl==2020.11.1.1
